// sdk/typescript/src/models/AgentStats.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AgentStats = {
    agentId?: string;
    totalEpochs?: number;
    totalVectors?: number;
    currentRoot?: string;
    lastEpoch?: number;
    createdAt?: string;
};